+++ connectome_manipulator/connectome_manipulation/manipulation/conn_rewiring.py
@@ -31,8 +31,9 @@ class ConnectomeRewiring(MorphologyCachingManipulation):
     The manipulation can be applied through the :func:`apply` method.
     """
 
-    # SONATA section type mapping: 0 = soma, 1 = axon, 2 = basal, 3 = apical
-    SEC_TYPE_MAP = {nm.AXON: 1, nm.BASAL_DENDRITE: 2, nm.APICAL_DENDRITE: 3}
+    # SONATA section type mapping (as in MorphIO): 1 = soma, 2 = axon, 3 = basal, 4 = apical
+    SEC_SOMA = 1
+    SEC_TYPE_MAP = {nm.AXON: 2, nm.BASAL_DENDRITE: 3, nm.APICAL_DENDRITE: 4}
 
     def __init__(self, nodes, writer, split_index=0, split_total=1):
         """Construct ConnectomeRewiring Manipulation and declare state vars..."""
@@ -904,7 +905,7 @@ class ConnectomeRewiring(MorphologyCachingManipulation):
         off_sel[sec_sel == -1] = 0.0  # Soma offsets must be zero
 
         # Synapse positions & (mapped) section types, computed from section & offset
-        type_sel = np.full_like(sec_sel, 0)
+        type_sel = np.full_like(sec_sel, SEC_SOMA)
         pos_sel = np.tile(morph.soma.center.astype(float), (len(sec_sel), 1))
         for idx in np.flatnonzero(sec_sel >= 0):
             type_sel[idx] = self.SEC_TYPE_MAP[morph.section(sec_sel[idx]).type]